fantasy_mma/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==8.0.0
httpx==0.26.0
redis==5.0.1
celery==5.3.6
supabase==2.3.1 